Fix output request of transfer and fix DAO tests
- Output Requests of transfer include now additional informations:
   - Centralized in AbstractTransfer
   - In addition: Id of the transfer, internal error code byte, original filename, final filename
- DAO tests: improve speed on database configuration using as much as possible memory instead of disk
- Improve reuse of Database DAO model from R66

+++ WaarpHttp/src/main/webapp/compute-hash.js
@@ -13,7 +13,7 @@ function uuidv4() {
 }
 
 function uuidlong() {
-  var array = new Uint32Array(1);
+  var array = new BigInt64Array(1);
   crypto.getRandomValues(array);
   return array[0];
 }
